refactor: moved app module to /modules dir
