+++ apps/cli/src/main.ts
@@ -1,5 +1,5 @@
 import { CommandFactory } from 'nest-commander';
-import { AppModule } from './app/app.module';
+import { AppModule } from './modules/app/app.module';
 
 const bootstrap = async () => {
   await CommandFactory.run(AppModule);
